polish the history framework
