+++ src/user_manage/service/history/add_history/AddingHistoryOutputBoundary.java
@@ -1,0 +1,4 @@
+package user_manage.service.history.add_history;
+
+public class AddingHistoryOutputBoundary {
+}
