+++ src/user_manage/service/history/read_history/ReadingHistoryInputBoundary.java
@@ -1,4 +1,4 @@
-package user_manage.service.history.service.read_history;
+package user_manage.service.history.read_history;
 
 public interface ReadingHistoryInputBoundary {
