+++ src/user_manage/service/history/read_history/ReadingHistoryInputData.java
@@ -1,0 +1,4 @@
+package user_manage.service.history.read_history;
+
+public class ReadingHistoryInputData {
+}
