+++ src/user_manage/service/history/read_history/interface_adpter/ReadingHistoryViewModel.java
@@ -1,0 +1,4 @@
+package user_manage.service.history.read_history.interface_adpter;
+
+public class ReadingHistoryViewModel {
+}
