--- src/user_manage/service/history/service/add_history/AddingHistoryInteractor.java
@@ -1,4 +1,0 @@
-package user_manage.service.history.service.add_history;
-
-public class AddingHistoryInteractor {
-}
