--- src/user_manage/service/history/service/add_history/Interface_adapter/AddingHistoryViewModel.java
@@ -1,4 +1,0 @@
-package user_manage.service.history.service.add_history.Interface_adapter;
-
-public class AddingHistoryViewModel {
-}
