--- src/user_manage/service/history/service/read_history/ReadingHistoryInteractor.java
@@ -1,4 +1,0 @@
-package user_manage.service.history.service.read_history;
-
-public class ReadingHistoryInteractor {
-}
