--- src/user_manage/service/history/service/read_history/ReadingHistoryOutputBoundary.java
@@ -1,4 +1,0 @@
-package user_manage.service.history.service.read_history;
-
-public interface ReadingHistoryOutputBoundary {
-}
